refactor: join all calculation logic into the single module
